:sparkles: DynamicTyper class
