+++ snk/cli/cli.py
@@ -15,6 +15,8 @@ from rich.console import Console
 from rich.syntax import Syntax
 from art import text2art
 
+from snk.cli.dynamic_typer import DynamicTyper
+
 
 from .config import (
     SnkConfig,
@@ -25,9 +27,9 @@ from .utils import add_dynamic_options, build_dynamic_cli_options, parse_config_
 from snk.pipeline import Pipeline
 
 
-class CLI:
+class CLI(DynamicTyper):
     """
-    Constructor for the CLI class.
+    Constructor for the dynamic Snk CLI class.
     Args:
       pipeline_dir_path (Path): Path to the pipeline directory.
     Side Effects:
@@ -122,40 +124,6 @@ class CLI:
                 "help_option_names": ["-h", "--help"],
             },
         )
-
-    def __call__(self):
-        """
-        Invoke the CLI.
-        Side Effects:
-          Invokes the CLI.
-        Examples:
-          >>> CLI(Path('/path/to/pipeline'))()
-        """
-        self.app()
-
-    def register_command(self, command: Callable, **command_kwargs) -> None:
-        """
-        Register a command to the CLI.
-        Args:
-          command (Callable): The command to register.
-        Side Effects:
-          Registers the command to the CLI.
-        Examples:
-          >>> CLI.register_command(my_command)
-        """
-        self.app.command(**command_kwargs)(command)
-
-    def register_callback(self, command: Callable, **command_kwargs) -> None:
-        """
-        Register a callback to the CLI.
-        Args:
-          command (Callable): The callback to register.
-        Side Effects:
-          Registers the callback to the CLI.
-        Examples:
-          >>> CLI.register_callback(my_callback)
-        """
-        self.app.callback(**command_kwargs)(command)
 
     def create_logo(self, tagline="A Snakemake pipeline CLI generated with snk", font="small"):
         """
@@ -279,9 +247,6 @@ class CLI:
                             fg=typer.colors.YELLOW,
                         )
                     remove_resource(copied_resource)
-    def error(self, msg):
-        typer.secho(msg, fg="red")
-        raise typer.Exit(1)
     
     def run(
         self,
